Fix ghost role cast
